% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNassAPI_funs.R
\name{getNass}
\alias{getNass}
\title{Extract NASS data from API by custom query components}
\usage{
getNass(apiKey, params)
}
\arguments{
\item{apiKey}{Personal key; Obtained from http://quickstats.nass.usda.gov/api}

\item{params}{query string and values}
}
\description{
Maximum flexibility function to download specified NASS data based on a user supplied
list of parameters. Requires knowledge of Quickstats API parameter codes
}
