% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNassAPI_funs.R
\name{getNassCounty_fc}
\alias{getNassCounty_fc}
\title{Extract NASS data by county from API - the whole shebang}
\usage{
getNassCounty_fc(apiKey, program = "CENSUS", sector = "CROPS",
  group = "FIELD CROPS", state, county)
}
\arguments{
\item{apiKey}{Personal key; Obtained from http://quickstats.nass.usda.gov/api}

\item{program}{'CENSUS' or 'SURVEY'}

\item{sector}{ie 'CROPS'}

\item{group}{ie 'FIELD CROPS'}

\item{state}{state 2 digit fips}

\item{county}{county 3-digit ANSI code}
}
\description{
Downloads NASS crop data by requested county, program source. Intended to retrieve 
all of the field crop data for a county, to be parsed by user.
}
\keyword{NASS}
\keyword{county}
