% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNassAPI_funs.R
\name{getNassCounty_short}
\alias{getNassCounty_short}
\title{Extract NASS data by county from API}
\usage{
getNassCounty_short(apiKey, program, shortDesc, state, county)
}
\arguments{
\item{apiKey}{Personal key; Obtained from http://quickstats.nass.usda.gov/api}

\item{program}{'CENSUS' or 'SURVEY'}

\item{shortDesc}{short description of desired data field; aka 'Data Item'}

\item{state}{state 2 digit fips}

\item{county}{county 3-digit ANSI code}
}
\description{
Downloads NASS crop data by requested county, program source, and short desc
}
\examples{


# set up variables for data calls
apiKey <- '28EAA9E6-8062-34A4-981A-B2ED4692228A' # my code, modified
program = 'CENSUS'
shortDesc <- 'AG LAND, CROPLAND - ACRES'
state <- '08'
county <- '001'

# simple example
oneCounty <- getNassCounty(apiKey, program, shortDesc, state, county)

# data frame of states and counties:
# make empty list to populate, and loop through counties
nass.list <- list()
for(i in 1:nrow(counties)){
  nass.list[[i]] <- getNASS(apiKey, program, sector, group, shortDesc, aggregation, 
                                state = counties$abb[i],
                                county = counties$COUNTYFP[i])
}
# convert list of data frames to 1 giant dataframe
nass.df <- do.call("rbind",nass.list) 

}
\keyword{NASS}
\keyword{county}
