% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNassAPI_funs.R
\name{getStateFips}
\alias{getStateFips}
\title{State Fips Lookup}
\usage{
getStateFips(statesWanted)
}
\arguments{
\item{statesWanted}{data frame with a column either 'StateName' or 'StateAbbrev' for states of interest}
}
\description{
joins table of Fips by either stateName or stateAbbrev. requires dplyr.
}
\examples{

}
