% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xdb_helper_funs.R
\name{makeExperimentTable}
\alias{makeExperimentTable}
\title{Make Experiment Master Table}
\usage{
makeExperimentTable(runTitle, ExpID, mukey, wthID, ExpCode, startyear, Nyears,
  startDOY, state = "KS", cropfp = "cropsn29Dec2016.cdb.xml")
}
\arguments{
\item{runTitle}{Name of overall run scenario}

\item{ExpID}{Sequential ID for experiment}

\item{mukey}{soil mukey from SSURGO}

\item{wthID}{weather station ID}

\item{ExpCode}{text string denoting sequence of annual climate-soil-crop-irrigation rotations, separated by '_'}

\item{startyear}{year to start experiment}

\item{Nyears}{Number of years in experiment}

\item{startDOY}{Starting day of year, fills in SDOY parameter}

\item{state}{state, 2-letter abbreviation. defaults to 'KS'}

\item{cropfp}{Filepath to crop .cdb.xml database file. Defaults to 'cropsn29Dec2016.cdb.xml'}
}
\description{
This function extracts Experiment level parameters from experiment combination
strings that have been parsed into soil and weather parameters. Inputs are
vectors of the same length, or single values to be repeated for all experiments.
Outputs a data frame containing arguments needed for `write_xdb_experiment`.
See http://salusmodel.glg.msu.edu/salus.ddb.xml for more
information about SALUS parameter options.
}
\details{
All other SALUS Experiment parameters are set currently as default.
}
\examples{
# Prepare the input using output from 02.21, GetExperiments
# Essentially df of ExpID and continuous management string
lemaExps <- lemaExps \%>\%
  mutate(nldas = str_sub(ExpCode, start = 10, end = 13),
        state = str_sub(ExpCode, start = 1, end = 3),
        mukey = str_sub(ExpCode, start = 3, end = 9),
        static = str_sub(ExpCode, start = 1, end = 13))

# translate mukey, expid, and nldas info into data.frame of Experiment parameters
exp_master <- makeExperimentTable('run title', lemaExps$ExpID, lemaExps$mukey,
                                  lemaExps$nldas, lemaExps$ExpCode, 2005, 11, 265)
}
\keyword{experiment}
\keyword{parameters}
\keyword{preparation}
\keyword{to}
\keyword{write}
