% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xdb_helper_funs.R
\name{parseRotationStrings}
\alias{parseRotationStrings}
\title{Parse Crop Rotations from Experiment String Code}
\usage{
parseRotationStrings(ExpCode, startyear, cropkey)
}
\arguments{
\item{ExpCode}{One text string denoting sequence of annual climate-soil-crop-irrigation rotations, separated by '_'}

\item{startyear}{year to start experiment}

\item{cropkey}{Look up table for CDL crop codes. column names should be 'cropCode' and 'crop'}
}
\description{
This function extracts rotation and management level parameters for
one experiment. Outputs a data frame containing year, crop,
and irrigation status, along with nldas and soil codes.
}
\details{
For any grid cells that include 'grassland' in the middle of a crop sequence,
this is assumed to be misclassified by the CDL and is changed to 'fallow'
}
\examples{
# make inputs
expCode <- '20266874911230010_20266874911230010_20266874911230010_20266874911231760_20266874911230010_20266874911230610_20266874911230010_20266874911230040_20266874911230240_20266874911230240_20266874911230611'
sYear <- 2006
cropkey <- data.frame(cropCode = c('001','004','005','024','061','176'),
                      crop = c('CORN','SORGHUM','SOYBEANS','WHEAT',
                                   'FALLOW','GRASS'))

rotations <- parseRotationStrings(expCode, sYear, cropkey)
}
\keyword{parameters}
\keyword{preparation}
\keyword{rotation}
\keyword{to}
\keyword{write}
