% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_processing.R
\name{salusRawToAnnual}
\alias{salusRawToAnnual}
\title{Process Raw SALUS Output}
\usage{
salusRawToAnnual(scratchDir, outDir, runname, startYear)
}
\arguments{
\item{scratchDir}{Raw SALUS output directory}

\item{outDir}{directory to export cleaned/summarized output}

\item{runname}{Run name}

\item{startYear}{First year of harvest}
}
\description{
This function takes the output daily files and converts daily sums to
annual values. Requires dplyr.
}
\examples{
# test
runResults <- salusRawToAnnual(scratchDir, outDir, runname, startYear)
}
\keyword{raw}
\keyword{results}
