% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_spatialize.R
\name{spatialize}
\alias{spatialize}
\title{Spatialize SALUS Results}
\usage{
spatialize(expt_df, variable, year, gridkey, template, aoi)
}
\arguments{
\item{expt_df}{Data.frame containing SALUS annual results in long format, loaded with readr::read_csv}

\item{variable}{Variable to process}

\item{year}{Year to process}

\item{gridkey}{data.frame key linking ExpID and template gridcell}

\item{template}{raster template}

\item{aoi}{Area of interest to clip to defined by a spdf}
}
\description{
Takes the processed annual results from SALUS and turns into a spatial raster,
based on the Experiment IDs - EXpID - and gridcell keys. Requires tidyverse,
raster, and rgdal
}
\examples{
# load formatted results output
library(readr)
resultsDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/salus/experiments/LEMA_historical_autoIrrigation_v02/results'
runname <- 'lema_historic_auto_v02'
expt_df <- read_csv(paste0(resultsDir, '/', runname, '.csv'))

# template raster
library(raster)
aeaProj <- "+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=23 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs"
template <- raster(nrow = 4359, ncol = 7108, crs = aeaProj,
                   xmn = -522870, xmx = -309630, ymn = 1777860,
                   ymx = 1908630)
template[] <- 1:ncell(template)

# variable wanted
variable <- 'IRRC_mm'

year <- 2006

# load my grid key from r data...
comboDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/salus/combinationGrids/spatialFilters'
load(file = paste0(comboDir,'/uniqueExperiments_gmd4p_Sheridan6_top6.RData'))
gridkey <- gridExps

rasterData <- spatialize(expt_df, variable, year, gridkey, template, aoi)
}
\keyword{results}
\keyword{spatialize}
