% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_spatialize_grasslandOverride.R
\name{spatializeAll_grass}
\alias{spatializeAll_grass}
\title{Spatialize wrapper: apply to multiple years}
\usage{
spatializeAll_grass(expt.df, variable, yearsWanted, gridkey, template, aoi,
  writeOut = "N", outFolder = "")
}
\arguments{
\item{variable}{Variable to process}

\item{yearsWanted}{vector of years to process}

\item{gridkey}{data.frame key linking ExpID and template gridcell}

\item{template}{raster template}

\item{aoi}{Area of interest to clip to defined by a spdf}

\item{writeOut}{Y or N}

\item{outFolder}{Needed if writeOut = 'Y'}

\item{expt_df}{Data.frame containing SALUS annual results in long format, loaded with readr::read_csv}
}
\description{
Function to apply the spatialize function to all years, and export rasters, optionally
}
\examples{
#' # example forthcoming
}
\keyword{all}
\keyword{results}
\keyword{spatialize}
\keyword{years}
