% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_spatialize.R
\name{summarizeWaterStacks}
\alias{summarizeWaterStacks}
\title{Annual Water Stats from Spatialized Results}
\usage{
summarizeWaterStacks(varStack, varName)
}
\arguments{
\item{varStack}{stack of annual rasters for variable of interest}

\item{varName}{name of variable}
}
\description{
Function to extract annual total and summary stats from stacks of annual
maps for the specified variable. Returns a data frame with annual totals and
per area stats for all years for the specified variable and irrigation status
}
\examples{
# example forthcoming
}
\keyword{results}
\keyword{summarize}
\keyword{water}
