% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_spatialize.R
\name{summarizeYieldStacks}
\alias{summarizeYieldStacks}
\title{Annual Yield Stats from Spatialized Results}
\usage{
summarizeYieldStacks(cropCode, yieldList, irrigationStatus)
}
\arguments{
\item{cropCode}{code for crop desired: MZ, WH, SG, SB, AL}

\item{yieldList}{List of stacks - 1 stack for each crop which includes all years}

\item{irrigationStatus}{Y or N - needs to be manually matched in input yield list}
}
\description{
Function to extract annual total and summary stats from stacks of annual
maps for the specified variable. Returns a data frame with annual totals and
per area stats for all years for the specified variable and irrigation status
}
\examples{
# example forthcoming
}
\keyword{results}
\keyword{summarize}
\keyword{yields}
