% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_batch_funs.R
\name{write_HPC_bat}
\alias{write_HPC_bat}
\title{Make HPCC Batch Files to run all experiments in run}
\usage{
write_HPC_bat(sh_vector, outFile)
}
\arguments{
\item{sh_vector}{A vector of sh shellscripts written by}

\item{outFile}{Path and filename with which to save the output bat file}
}
\description{
This function makes a bash script to execute queue submissions to MSU's High
Performance Computing Cluster (HPCC).
}
\examples{
# function arguments
shNames <- paste0('test_',1:10,'.sh')
fileOut <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/SALUS/testHPCCrun/test.bat'

# write out .bat file
write_HPC_bat(shNames, fileOut)
}
\keyword{HPCC}
\keyword{bat}
\keyword{batch}
\keyword{preparation}
