% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_batch_funs.R
\name{write_HPC_shell}
\alias{write_HPC_shell}
\title{Make HPCC Shell Files to Run Xdbs}
\usage{
write_HPC_shell(shDir, hpcHomeDir, hpcXdbDir, hpcOutDir, xdb, sdb, wdbZip,
  DayVars, SeaVars, walltime, memory, cdb)
}
\arguments{
\item{shDir}{Directory in which to save the output shell sh file. the sh filename is pulled from xdb filename}

\item{hpcHomeDir}{Directory on HPCC in which source files reside for this run (sdb, xdb, zipped weather, etc). Should end in a forward slash}

\item{hpcOutDir}{Directory on HPCC to write results to, Should end in a forward slash}

\item{xdb}{filename of the experiment file, without extension. sh name will be pulled from this.}

\item{sdb}{filename of the soil file (no path, no .sdb.xml extension)}

\item{wdbZip}{filename of the zipped weather file. Script will unzip on the node}

\item{DayVars}{CSV string of SALUS daily variables to return}

\item{SeaVars}{CSV string of SALUS seasonal variables to return}

\item{walltime}{For HPCC - how long you expect the job to run. Format HH:MM:SS}

\item{memory}{For HPCC - how much memory the job will need. ie, '2gb' or '2000mb'}

\item{cdb}{filename of the crop cdb.xml file, without the extension}
}
\description{
This function makes a shell script to defining experiment runs for MSU's High
Performance Computing Cluster (HPCC). This bash script copies the files needed
to run SALUS (salus_gnu, gdb, cdb, sdb, xdb, and wdb files) to the current
directory. Then the bash script will run SALUS. Then the bash script will remove
the SALUS files from the current directories.Then the bash script will copy
the SALUS results to your scratch directory. Modified from python writeTorque
function by Brian Baer and Lydia Rill.
}
\details{
Recommended to follow this with the `write_HPCC_bat` function in order to
generate the batch bat file that will submit multiple shell files to the HPCC queue.
}
\examples{
#set variables
shDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/SALUS/testHPCCrun'
hpcHomeDir <- '/mnt/home/deinesji/Example_SALUS_wheat/'
hpcOutDir <- '/mnt/home/deinesji/Example_SALUS_wheat/results/'
xdb <- 'lema_wheat_continuous'
sdb <- "KS"
wdbZip <- 'testHPCCrun.tar.gz'
DayVars <- 'ExpID,Title,SpeciesID,GWAD,IRRC,CWAD,DRNC,PREC,LAI'
SeaVars <- 'ExpID,Title,SpeciesID,GWAD,IRRC,CWAD,DRNC,PREC'
walltime <- '01:00:00'
memory <- '2000mb'

write_HPC_shell(shDir, hpcHomeDir, hpcOutDir, xdb, sdb, wdbZip, DayVars,
                SeaVars, walltime, memory)
}
\keyword{HPCC}
\keyword{preparation}
\keyword{sh}
\keyword{shell}
