% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_runScripts.R
\name{write_HPC_shell_R}
\alias{write_HPC_shell_R}
\title{Make HPCC Shell Files to Run R Scripts for Prep and Run Summaries}
\usage{
write_HPC_shell_R(outDir, runname, scriptTypeName, hpcBaseDir, scriptName,
  walltime, memory)
}
\arguments{
\item{outDir}{Directory in which to save the output shell sh file. the sh filename is pulled from run filename}

\item{runname}{Run name for the set of files}

\item{scriptTypeName}{appended to filename}

\item{hpcBaseDir}{Directory on HPCC within which are subdirectories for each run}

\item{scriptName}{Name of R script to run}

\item{walltime}{For HPCC - how long you expect the job to run. Format HH:MM:SS}

\item{memory}{For HPCC - how much memory the job will need. ie, '2gb' or '2000mb'}
}
\description{
This function makes a shell script to execute the 03.00_salusXdb_experimentWriter
script for MSU's High Performance Computing Cluster -HPCC-. This bash script
loads the proper version of R, sets the R library path, and executes the r file.
The R file -experiment writer- writes the salus xdb.xml's for the run based on
the run parameter file and the set of unique experiments.
}
\examples{
#set variables
}
\keyword{HPCC}
\keyword{preparation}
\keyword{sh}
\keyword{shell}
