% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xdb_funs.R
\name{write_xdb_mFertilize}
\alias{write_xdb_mFertilize}
\title{Write SALUS xdb.xml Fertlizer Management for Rotation Component}
\usage{
write_xdb_mFertilize(outFile, Year, DOY, ANfer, closeComponent = "N")
}
\arguments{
\item{outFile}{Full file path for an existing .xdb.xml to appended to, created with write_xdb_topMatter}

\item{Year}{Year of fertilization event - if needed (fertilizing on Reported date); XXXX}

\item{DOY}{day of year of fertilization event; only needed if fert management is 'R' or 'Y'. Valid range: 1-366}

\item{ANfer}{amount of nitrogen in fertilizer <- 78.47 # kg / ha}
}
\description{
This function writes the fertilizer management arguments after write_xdb_rotation has been run. See http://salusmodel.glg.msu.edu/salus.ddb.xml for more
information about SALUS parameter options. This function should be followed with additional
write_xdb_m* (rotation managements), and write_xdb_bottomMatter functions to complete the Experiment file.
Includes an option to close the rotation compoment if this is the last management specified.
}
\details{
All other SALUS fertilizer parameters are set currently as default.
}
\examples{
# file path to existing .xdb.xml file created by write_xdb_topMatter
fileOut <- 'C:/Users/deinesji/1PhdJill/test.xdb.xml'

# append rotation component parameters to .xdb.xml; numeric parameters can be character or numeric
write_xdb_mFertilize(outFile = fileOut, Year = 2007, DOY = 75, ANfer = 78.47)
}
\keyword{create}
\keyword{fertilizer}
\keyword{managment}
\keyword{xdb,}
