% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xdb_funs.R
\name{write_xdb_mFertilize_Auto}
\alias{write_xdb_mFertilize_Auto}
\title{Write SALUS xdb.xml AUTO Fertlizer Management for Rotation Component}
\usage{
write_xdb_mFertilize_Auto(outFile, SoilNC, SoilNX, DSoilN = "5",
  NCode = "FE010", closeComponent = "N")
}
\arguments{
\item{outFile}{Full file path for an existing .xdb.xml to appended to, created with write_xdb_topMatter}

\item{SoilNC}{Threshold, N stress factor, percent, required}

\item{SoilNX}{Amount per application, kg N per ha, required}

\item{DSoilN}{Application depth, cm, required}

\item{NCode}{Materian code, required}

\item{closeComponent}{'Y' or 'N' . Should the rotation component be closed? defaults to 'N'}
}
\description{
This function writes the fertilizer management arguments after write_xdb_rotation has been run. See http://salusmodel.glg.msu.edu/salus.ddb.xml for more
information about SALUS parameter options. This function should be followed with additional
write_xdb_m* (rotation managements), and write_xdb_bottomMatter functions to complete the Experiment file.
Includes an option to close the rotation compoment if this is the last management specified.
}
\details{
All other SALUS fertilizer parameters are set currently as default.
}
\examples{
#' # example forthcoming
}
\keyword{create}
\keyword{fertilizer}
\keyword{managment}
\keyword{xdb,}
