% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xdb_funs.R
\name{write_xdb_mHarvest_maturity}
\alias{write_xdb_mHarvest_maturity}
\title{Write SALUS xdb.xml Harvest Management for Rotation Component: At Maturity}
\usage{
write_xdb_mHarvest_maturity(outFile, HKnDnPc = 100, HBPc = 0,
  HPc = 100, closeComponent = "N")
}
\arguments{
\item{outFile}{Full file path for an existing .xdb.xml to appended to, created with write_xdb_topMatter}

\item{HKnDnPc}{Harvest knock-down percent between 0 and 100. Defaults to 100. Percent of stalk/leaves knocked down at harvest.}

\item{HBPc}{Percent of byproduct harvested between 0 and 100. Defaults to 0. basically how much of the leaves/stem to you want to remove from the field}

\item{HPc}{Harvest percentage between 0 and 100. Defaults to 100.}

\item{closeComponent}{'Y' or 'N' . Should the rotation component be closed? defaults to 'N'}
}
\description{
This function writes the harvest management arguments after write_xdb_rotation has been run. See http://salusmodel.glg.msu.edu/salus.ddb.xml for more
information about SALUS parameter options. This function should be followed with additional
write_xdb_m* (rotation managements), and write_xdb_bottomMatter functions to complete the Experiment file.
Includes an option to close the rotation compoment if this is the last management specified.
}
\details{
All other SALUS fertilizer parameters are set currently as default.
}
\examples{
# file path to existing .xdb.xml file created by write_xdb_topMatter
fileOut <- 'C:/Users/deinesji/1PhdJill/test.xdb.xml'

# append rotation component parameters to .xdb.xml; numeric parameters can be character or numeric
write_xdb_mHarvest(outFile = fileOut)
}
\keyword{create}
\keyword{fertilizer}
\keyword{managment}
\keyword{xdb,}
