% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xdb_funs.R
\name{write_xdb_mIrrigate_Auto}
\alias{write_xdb_mIrrigate_Auto}
\title{Write SALUS xdb.xml Irrigation Management for Rotation Component: Auto}
\usage{
write_xdb_mIrrigate_Auto(outFile, AIrAm = "", DSoil = 20, ThetaC = 50,
  IAMe = "IR004", closeComponent = "N")
}
\arguments{
\item{outFile}{Full file path for an existing .xdb.xml to appended to, created with write_xdb_topMatter}

\item{AIrAm}{Amount per irrigation if fixed in mm; defaults to "" empty quotes}

\item{DSoil}{Management depth for irrig. cm. Defaults to 20 cm}

\item{ThetaC}{Threshold for automatic appl.- percent of max. available w. Defaults to 50}

\item{IAMe}{Method for automatic appl. defaults to 'IR004' for sprinkler}

\item{closeComponent}{'Y' or 'N' . Should the rotation component be closed? defaults to 'N'}
}
\description{
This function writes the irrigation management arguments after write_xdb_rotation has been run.
See http://salusmodel.glg.msu.edu/salus.ddb.xml for more
information about SALUS parameter options. This function should be followed with additional
write_xdb_m* (rotation managements), and write_xdb_bottomMatter functions to complete the Experiment file.
Includes an option to close the rotation compoment if this is the last management specified.
}
\details{
All other SALUS tillage parameters are set currently as default.
}
\examples{
# file path to existing .xdb.xml file created by write_xdb_topMatter
fileOut <- 'C:/Users/deinesji/1PhdJill/test.xdb.xml'

# append rotation component parameters to .xdb.xml; numeric parameters can be character or numeric
write_xdb_mTillage(outFile = fileOut)
}
\keyword{create}
\keyword{managment}
\keyword{tillage}
\keyword{xdb,}
