% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xdb_funs.R
\name{write_xdb_mPlanting}
\alias{write_xdb_mPlanting}
\title{Write SALUS xdb.xml Planting Management for Rotation Component}
\usage{
write_xdb_mPlanting(outFile, CropMod, SpeciesID, cultivar, Year, DOY, Ppop,
  RowSpc, closeComponent = "N")
}
\arguments{
\item{outFile}{Full file path for an existing .xdb.xml to appended to, created with write_xdb_topMatter}

\item{CropMod}{Crop model to be used. C = complex, S = simple. Intermediate I still under development}

\item{SpeciesID}{Species ID}

\item{Year}{Year of planting - if needed -planting on Reported date; XXXX}

\item{DOY}{day of year of planting; only needed if planting management is 'R' or 'Y'. Valid range: 1-366}

\item{Ppop}{Plant population at seeding; plants m^-2}

\item{RowSpc}{Row spacing, cm}

\item{closeComponent}{'Y' or 'N' . Should the rotation component be closed? defaults to 'N'}

\item{CultivarID}{cultivar ID, only needed for complex crop model}
}
\description{
This function writes the planting management arguments after write_xdb_rotation has been run. See http://salusmodel.glg.msu.edu/salus.ddb.xml for more
information about SALUS parameter options. This function should be followed with additional
write_xdb_m* (rotation managements), and write_xdb_bottomMatter functions to complete the Experiment file.
Includes an option to close the rotation compoment if this is the last management specified.
}
\details{
All other SALUS planting parameters are set currently as default.
}
\examples{
# file path to existing .xdb.xml file created by write_xdb_topMatter
fileOut <- 'C:/Users/deinesji/1PhdJill/test.xdb.xml'

# append rotation component parameters to .xdb.xml; numeric parameters can be character or numeric
write_xdb_mPlanting(outFile = fileOut, CropMod = 'C', SpeciesID = 'WH', CultivarID = 'P25R40',
                   Year = 2006, DOY = 276, Ppop = 494.21, RowSpc = 19.05)
}
\keyword{create}
\keyword{managment}
\keyword{planting}
\keyword{xdb,}
