% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xdb_funs.R
\name{write_xdb_rotation}
\alias{write_xdb_rotation}
\title{Write SALUS xdb.xml Rotation Component Level Parameters}
\usage{
write_xdb_rotation(outFile, OrderNum, Title, IIrrI, IferI, ITilI, IHarI,
  IResI)
}
\arguments{
\item{outFile}{Full file path for an existing .xdb.xml to appended to, created with write_xdb_topMatter}

\item{OrderNum}{Order of the treatment. Minimum = 1. 1 will initialize the Rotation_components section}

\item{Title}{Rotation component title, such as 'maize', 'Wheat'}

\item{IIrrI}{Irrigation management. 'N' - not irrigated; 'A' - auto profile refill; 'F' - auto with fixed amount; 'R' - reported date, 'Y' - reported doy, 'D' - days after planting}

\item{IferI}{Fertilizer management. 'N' - not irrigated; 'A' - automatic; 'F' - auto with fixed amount; 'R' - reported date, 'Y' - reported doy, 'D' - days after planting}

\item{ITilI}{Tillage management. 'N' - no tillage; 'A' - automatic; 'R' - reported; 'Y' - reported doy, 'D' - days after planting}

\item{IHarI}{Harvest management. A: Automatic; D: Days after Planting; G: At Growth Stages; M: At Maturity R: Reported Date W: Harvest when crop reaches given weight Y: Reported Day of Year}

\item{IResI}{Residue management}
}
\description{
This function writes the rotation component arguments. Subsequent write_xdb_m* functions append information
about management options. See http://salusmodel.glg.msu.edu/salus.ddb.xml for more
information about SALUS parameter options. This function should be followed with
write_xdb_m* (rotation management), and write_xdb_bottomMatter functions to complete the Experiment file.
}
\details{
All other SALUS Rotation Component parameters are set currently as default.
}
\examples{
# file path to existing .xdb.xml file created by write_xdb_topMatter
fileOut <- 'C:/Users/deinesji/1PhdJill/test.xdb.xml'

# append rotation component parameters to .xdb.xml; numeric parameters can be character or numeric
write_xdb_rotation(outFile = fileOut, OrderNum = 1, Title = 'Wheat', IIrrI = 'N',
                   IferI = 'R', ITilI = 'R', 'IHarI' = 'R', IResI = 'A')
}
\keyword{create}
\keyword{xdb}
