% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xdb_funs.R
\name{write_xdb_topMatter}
\alias{write_xdb_topMatter}
\title{Write SALUS xdb.xml Top Matter}
\usage{
write_xdb_topMatter(outFile)
}
\arguments{
\item{outFile}{Full file path for the .xdb.xml to create}
}
\description{
This function initializes a SALUS experiment file by creating the file and
initializing the <XDB> tag. Subsequent write_xdb_* functions append information
about the experiments and rotational components.
}
\examples{
# desired file path and file name for the output to be created
fileOut <- 'C:/Users/deinesji/1PhdJill/test.xdb.xml'
# create output
write_xdb_topMatter(fileOut)
}
\keyword{create}
\keyword{xdb}
